+++ atlasdb-client/src/main/java/com/palantir/atlasdb/AtlasDbConstants.java
@@ -109,4 +109,6 @@ public class AtlasDbConstants {
 
     public static final int CASSANDRA_TABLE_NAME_CHAR_LIMIT = 48;
     public static final int POSTGRES_TABLE_NAME_CHAR_LIMIT = 63;
+
+    public static final String SCHEMA_V2_TABLE_NAME = "V2Table";
 }
